fix: Phase17 - DF-3スコア計算、FK-1属性吸収クランプ、FY-1建設素材消費、CP-3除算ガード追加
Agent-Logs-Url: https://github.com/maki04591128-tech/game/sessions/b771672a-cbdb-4de8-8ba9-6ce5f526638a

+++ src/RougelikeGame.Core/Systems/GameClearSystem.cs
@@ -67,7 +67,7 @@ public static class GameClearSystem
     public static ClearScore CalculateScore(int totalTurns, int totalDeaths, int highestLevel, int maxFloor)
     {
         // ターンボーナス（少ないほど高得点）
-        int turnBonus = Math.Max(0, 10000 - totalTurns / 5);
+        int turnBonus = Math.Max(0, 10000 - (int)((double)totalTurns / 5.0));
         // 死亡ペナルティ
         int deathPenalty = totalDeaths * 500;
         // レベルボーナス
